# ui/desktop/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "aster-server",
    "description": "An AI agent",
    "contact": {
      "name": "Block",
      "email": "ai-oss-tools@block.xyz"
    },
    "license": {
      "name": "Apache-2.0"
    },
    "version": "1.19.0"
  },
  "paths": {
...
    "/agent/add_extension": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "agent_add_extension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddExtensionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Extension added",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/call_tool": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "call_tool",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CallToolRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resource read successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallToolResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "404": {
            "description": "Resource not found"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/read_resource": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "read_resource",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReadResourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Resource read successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadResourceResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "404": {
            "description": "Resource not found"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/remove_extension": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "agent_remove_extension",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveExtensionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Extension removed",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/resume": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "resume_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResumeAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - invalid working directory"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/start": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "start_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartAgentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Agent started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/agent/tools": {
      "get": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "get_tools",
        "parameters": [
          {
            "name": "extension_name",
            "in": "query",
            "description": "Optional extension name to filter tools",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "session_id",
            "in": "query",
            "description": "Required session ID to scope tools to a specific session",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Tools retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ToolInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/agent/update_from_session": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "update_from_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFromSessionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update agent from session data successfully"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          }
        }
      }
    },
    "/agent/update_provider": {
      "post": {
        "tags": [
          "super::routes::agent"
        ],
        "operationId": "update_agent_provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Provider updated successfully"
          },
          "400": {
            "description": "Bad request - missing or invalid parameters"
          },
          "401": {
            "description": "Unauthorized - invalid secret key"
          },
          "424": {
            "description": "Agent not initialized"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/config/extensions/{name}": {
      "delete": {
        "tags": [
          "super::routes::config_management"
        ],
        "operationId": "remove_extension",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Extension removed successfully",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Extension not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddExtensionRequest": {
        "type": "object",
        "required": [
          "session_id",
          "config"
        ],
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ExtensionConfig"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
...
      "CallToolRequest": {
        "type": "object",
        "required": [
          "session_id",
          "name",
          "arguments"
        ],
        "properties": {
          "arguments": {},
          "name": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        }
      },
      "CallToolResponse": {
        "type": "object",
        "required": [
          "content",
          "is_error"
        ],
        "properties": {
          "_meta": {
            "nullable": true
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Content"
            }
          },
          "is_error": {
            "type": "boolean"
          },
          "structured_content": {
            "nullable": true
          }
        }
      },
...
      "ReadResourceRequest": {
        "type": "object",
        "required": [
          "session_id",
          "extension_name",
          "uri"
        ],
        "properties": {
          "extension_name": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          },
          "uri": {
            "type": "string"
          }
        }
      },
...
    }
  }
}